Training/requirements-training.txt
pyyaml
numpy
torch
pillow
torchvision
open_clip_torch
mlflow

psycopg2-binary
python-swiftclient
python-keystoneclient

tqdm
requests